pyproject.toml
[project]
name = "iata-teletype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.4",
    "google-cloud-pubsub>=2.34.0",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "pytest>=8.3.5",
    "textual>=6.2.1",
    "uvicorn>=0.33.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[project.scripts]
teletype-api = "message_builder.api:run"
teletype-tui = "tui.app:run"
